implement no workflows state
